UBUNTU: SAUCE: drm/dp: drm_edp_backlight_set_level: do not always send 3-byte commands
BugLink: https://bugs.launchpad.net/bugs/2121477

At least some panels using the LSB register are not happy with the
unconditional increase of the command buffer to 3 bytes.

With the BOE NE14QDM in my Dell Latitude 7455, the recent patches for
luminance based brightness have introduced a regression: the brightness
range stopped being contiguous and became nonsensical (it probably was
interpreting the last 2 bytes of the buffer and not the first 2).

Change from using a fixed sizeof() to a length variable that's only
set to 3 when luminance is used. Let's leave the default as 2 even for
the single-byte version, since that's how it worked before.

Fixes: f2db78e37fe7 ("drm/dp: Modify drm_edp_backlight_set_level")
Signed-off-by: Val Packett <val@packett.cool>
Reviewed-by: Suraj Kandpal <suraj.kandpal@intel.com>
Link: https://lore.kernel.org/r/20250706204446.8918-1-val@packett.cool
Signed-off-by: Abel Vesa <abel.vesa@linaro.org>
(cherry picked from commit d731952f09071bfa4734b78f10c76a2defcf6231 gitlab.com/Linaro/arm64-laptops/linux)
Signed-off-by: Tobias Heider <tobias.heider@canonical.com>